!!! image-android: Adds support for FDT in the qcom format.
 * https://github.com/osm0sis/mkbootimg/blob/2017.12.13/bootimg.h#L46

This is a hack, mainly because my pipeline in my personal project uses
that specific `mkbootimg`. I didn't know before starting that this was a
qualcomm specific variant of android boot images.

Once this is verified as working, it should be possible to cleanup and
port the actual upstream android FDT loading.

 * https://android.googlesource.com/platform/system/core/+/master/mkbootimg/include/bootimg/bootimg.h#1