drivers/video/rockchip/rk_edp.c: Add rk3399 support
According to linux commit 82872e42bb1501dd9e60ca430f4bae45a469aa64,
rk3288 and rk3399 eDP IPs are nearly the same, the difference is in the grf register
(SOC_CON6 versus SOC_CON20). So, change the code to use the right
register on each IP.

The clocks don't seem to be the same, the eDP clock is not at index 1 on rk3399,
so don't try changing the clock at index 1 to rate 0 on rk399. Also, enable all
clocks, in case it's needed.

Signed-off-by: Arnaud Patard <arnaud.patard@rtp-net.org>

http://people.hupstream.com/~rtp/pbp/rk_edp_rk3399.patch