configs: Move usb_pgood_delay workaround to board config
Instead of adding the workaround in the configuration, let's put
intrinsic known truths about the platform in a file which is guaranteed
to be built.

This way, the usb_pgood_delay workarounds is *always* present in the
default environment, and clobbering of PREBOOT will not accidentally
break the workaround.

Furthermore, this enables moving the USB_KEYBOARD usb start out of the
config files, as previously this workaround would be applied after the
intrinsic usb start.