[HACK] Somehow fix distroboot not changing to SD
This change is equivalent to another RK3588 vendor change
(4c20c87da731fefb275b17fe6c2655f7f5d94134), but less intrusive.

It has not been investigated why `mmc dev 1` does not work until `mmc
list` has been called at least once. It must not be specific to `mmc
list`, but whatever is the cause, `mmc list` "unclogs" it.

My assumption is that when SPL is loaded from mmc or SD, it does
something that isn't being done when it is loaded from SPI. This
is assumed to be the reason why it works when not loaded from SPI.