efi_loader: Disable EFI reset for RK3399
Most (all?) RK3399 in U-Boot don't support EFI reset.

Having it enabled means reboot is broken in the operating system.

A better solution would be to better expose a dependency such that
boards or SoCs or architectures can choose whether this function is used
or not.

Another better(?) solution would be to expose this in U-Boot DT entries,
and defaulting it to true as needed.

The actual solution would be to implement reset for all boards that
don't have it.

Signed-off-by: Samuel Dionne-Riel <samuel@dionne-riel.com>