WIP: spl: Rework SPL_RAM support to be more generic
The main reason behind this work, is to try to straighten out the
spaghetti that the function had become.

The function was used to uphold different load methods that are not
actually 100% related, other than by the fact that they use data found
in RAM. A tenuous link, since the next stage is highly likely to be
found in RAM at some point during the lifetime of U-Boot.

The methods this upheld were (non-exhaustively):

 - Loading next stage from SPL image data
 - Loading next stage from DFU-loaded information
 - Loading next stage from binman referenced data
 - Loading next stage from data already in RAM

And using the data in these ways:

 - Using location as a FIT image
 - Using location as Legacy image

A nuance to consider here his that these methods each tried to use the
data in slightly divergent ways. For example loading the data from a
FIT image only was possible if the SPL itself is a FIT image, and from
the SPL itself.

This, instead, uses any plausible location, from the previous schemes,
and tries all of the different schemes at those locations.

With these changes, it is possible to use `binman` to build a non-FIT
SPL image (e.g. for Rockchip systems), which includes the unmodified
U-Boot proper FIT image, which is referenced through its binman symbol.

In this exact situation, allowing U-Boot proper to run entirely from
“maskrom” mode using `rkdeveloptool` or `rockusb`. No traces of U-Boot
to be found on any storage medium.